+++ experiments/long_runs/land_region.jl
@@ -78,15 +78,10 @@ function setup_model(FT, context, start_date, Δt, domain, toml_dict)
     )
 
     # Overwrite some defaults for the canopy model
-    # Energy model
-    energy_args = Canopy.BigLeafEnergyParameters(toml_dict)
-    energy = Canopy.BigLeafEnergyModel{FT}(energy_args)
-
     # Plant hydraulics
     retention_model = Canopy.PlantHydraulics.LinearRetentionCurve(toml_dict)
     hydraulics = Canopy.PlantHydraulicsModel{FT}(
         surface_domain,
-        LAI,
         toml_dict;
         retention_model,
     )
@@ -105,7 +100,6 @@ function setup_model(FT, context, start_date, Δt, domain, toml_dict)
         LAI,
         toml_dict;
         prognostic_land_components = (:canopy, :snow, :soil, :soilco2),
-        energy,
         hydraulics,
         z_0m,
         z_0b,